Crate spin_factors_executor

Structs

FactorsExecutor

A FactorsExecutor manages execution of a Spin app.

FactorsExecutorApp

A FactorsExecutorApp represents a loaded Spin app, ready for instantiation.

FactorsInstanceBuilder

A FactorsInstanceBuilder manages the instantiation of a Spin component instance.

InstanceState

InstanceState is the [spin_core::Store] data for an instance.

Traits

ComponentLoader

A ComponentLoader is responsible for loading Wasmtime Components.

ExecutorHooks